bluetooth-broadcom.bb: Rework bluetooth firmware loading
There are cases when the rootfs is read-only. In such cases,
the approach of updating the systemd service responsible for
loading the bluetooth firmware will not work.

Instead we rework it and use an udev rule and systemd service
combination to load the bluetooth firmware.

Also, we change the logic of getting the /dev node because
the current device tree does not define a node named "bt_uart"
and instead we grep for "wireless-bluetooth".

Changelog-entry: Fix bluetooth firmware loading
Signed-off-by: Florin Sarbu <florin@balena.io>