u-boot: Load uEnv.txt from the boot partition
Loading additional dtb(s) for the Radxa Zero can be done by listing
the required dtb(s) in uEnv.txt. We now have moved uEnv.txt to the
boot partition which is read-write as opposed to having it in the rootfs
which is read-only.

Changelog-entry: Load uEnv.txt from the boot partition for the Radxa Zero
Signed-off-by: Florin Sarbu <florin@balena.io>