u-boot-rockpi-4: Set the boot partition as bootable for flasher images
This is needed for flasher images that have the kernel, device tree,
extlinux.conf, overlays and hw_intfc.conf in the boot partition.
In addition, when we are not booting a flasher image we add as backup
the boot partition so that we have a working system in case of a HUP
rollback to an older OS version which only has extlinux.conf in the boot
partition and not in the root partition.

Changelog-entry: Set the boot partition as bootable for flasher images
Signed-off-by: Florin Sarbu <florin@balena.io>