rkwifibt-firmware.bb: Use nonarch_base_libdir variable to change install path to /usr/lib
Using the nonarch_base_libdir variable to change the install path to /usr/lib instead
of hardcoding directly to /lib will help us when we enable the usrmerge feature.

We add the entire recipe here from the existing BSP repo because https://github.com/radxa/meta-radxa.git
is now a public archive. We must delete this recipe from here when we migrate to the vendor recommended BSP
on the kirkstone branch.

Changelog-entry: Use nonarch_base_libdir to install files in /usr/lib instead of /lib for the rkwifibt-firmware recipe
Signed-off-by: Florin Sarbu <florin@balena.io>